requirements.txt
flask
Flask-SQLAlchemy
sqlalchemy
requests
flask-cors
google-cloud-secret-manager
pymysql
